+++ lib/mkmf/lite.rb
@@ -66,7 +66,7 @@ module Mkmf
     end
 
     # The version of the mkmf-lite library
-    MKMF_LITE_VERSION = '0.9.0'
+    MKMF_LITE_VERSION = '0.9.1'
 
     class << self
       def configuration